third_party/WebKit/LayoutTests/http/tests/devtools/compiler-script-mapping-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests SourceMap and CompilerScriptMapping.


Running: testCompilerScriptMapping
Adding compiled.js
compiled.js UISourceCode arrived
compiled.js script found
source1.js UISourceCode arrived
source2.js UISourceCode arrived
Location checks passed. Requesting content
Content1 arrived.
Content2 arrived.

Running: testInlinedSourceMap
Adding compiled.js
compiled.js UISourceCode arrived
compiled.js script found
source3.js UISourceCode arrived
Location checks passed. Requesting content
<source content> === <source content>

Running: testSourceMapCouldNotBeLoaded
Adding compiled.js
SourceMap Failed to load.
test://evaluations/3/compiler-script-mapping.js
